backend/requirements.txt
Django
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
django-cors-headers